Move node attributes to Dynamo.Nodes.
