+++ src/Libraries/Analysis/SurfaceData.cs
@@ -5,6 +5,8 @@ using System.Linq;
 using Autodesk.DesignScript.Geometry;
 using Autodesk.DesignScript.Runtime;
 using Dynamo.Models;
+using Dynamo.Nodes;
+
 namespace Analysis
 {
     /// <summary>
