+++ src/Libraries/CoreNodeModels/ColorRange.cs
@@ -5,6 +5,7 @@ using System.Linq;
 
 using DSCore;
 using Dynamo.Models;
+using Dynamo.Nodes;
 using DSCoreNodesUI.Properties;
 using ProtoCore.AST.AssociativeAST;
 using ProtoCore.Mirror;
