+++ src/Libraries/CoreNodeModels/String.cs
@@ -1,4 +1,5 @@
 ﻿using Dynamo.Models;
+using Dynamo.Nodes;
 using ProtoCore.AST.AssociativeAST;
 
 using System.Collections.Generic;
