+++ src/Libraries/CoreNodes/Files.cs
@@ -1,5 +1,6 @@
 ﻿using Autodesk.DesignScript.Runtime;
 using Dynamo.Models;
+using Dynamo.Nodes;
 using System;
 using System.Collections;
 using System.Collections.Generic;
