+++ test/DynamoCoreTests/NodeConstructionTests.cs
@@ -1,4 +1,5 @@
 ﻿using Dynamo.Models;
+using Dynamo.Nodes;
 
 using NUnit.Framework;
